+++ src/App/Fossa/Analyze.hs
@@ -87,6 +87,7 @@ import Strategy.NuGet.PackagesConfig qualified as PackagesConfig
 import Strategy.NuGet.Paket qualified as Paket
 import Strategy.NuGet.ProjectAssetsJson qualified as ProjectAssetsJson
 import Strategy.NuGet.ProjectJson qualified as ProjectJson
+import Strategy.Pub qualified as Pub
 import Strategy.Python.Pipenv qualified as Pipenv
 import Strategy.Python.Poetry qualified as Poetry
 import Strategy.Python.Setuptools qualified as Setuptools
@@ -195,6 +196,7 @@ discoverFuncs =
   , Poetry.discover
   , ProjectAssetsJson.discover
   , ProjectJson.discover
+  , Pub.discover
   , RPM.discover
   , Rebar3.discover
   , RepoManifest.discover
@@ -216,8 +218,8 @@ runDependencyAnalysis (BaseDir basedir) filters project =
     Nothing -> logInfo $ "Skipping " <> pretty (projectType project) <> " project at " <> viaShow (projectPath project) <> ": no filters matched"
     Just targets -> do
       logInfo $ "Analyzing " <> pretty (projectType project) <> " project at " <> pretty (toFilePath (projectPath project))
-      graphResult <- Diag.runDiagnosticsIO . stickyDiag $ projectDependencyGraph project targets
-      Diag.withResult SevWarn graphResult (output . mkResult project)
+      graphResult <- Diag.runDiagnosticsIO . stickyDiag $ projectDependencyResults project targets
+      Diag.withResult SevWarn graphResult (output . mkResult basedir project)
 
 applyFiltersToProject :: Path Abs Dir -> AllFilters -> DiscoveredProject n -> Maybe FoundTargets
 applyFiltersToProject basedir filters DiscoveredProject{..} =