+++ test/Maven/DepTreeSpec.hs
@@ -1,9 +1,21 @@
 module Maven.DepTreeSpec (spec) where
 
+import Data.Text (Text)
 import Data.Text.IO qualified as TextIO
-import Strategy.Maven.DepTree (DotGraph (..), PackageId (..), parseDotGraphs)
-import Test.Hspec (Spec, describe, it, runIO)
-import Test.Hspec.Megaparsec (shouldParse)
+import DepTypes (
+  DepEnvironment (..),
+  DepType (MavenType),
+  Dependency (..),
+  VerConstraint (..),
+ )
+import Strategy.Maven.DepTree (
+  DotGraph (..),
+  PackageId (..),
+  parseDotGraphs,
+  toDependency,
+ )
+import Test.Hspec (Spec, describe, it, runIO, shouldBe)
+import Test.Hspec.Megaparsec (shouldParse, shouldSucceedOn)
 import Text.Megaparsec (parse)
 
 -- TODO: Other tests:
@@ -21,6 +33,35 @@ spec =
     it "parses multiple dot graphs in one file" $
       parse parseDotGraphs fixtureMultiFile multi `shouldParse` fixtureMultiGraph
 
+    acme <- runIO $ TextIO.readFile fixtureAcmeFile
+    it "parses real-world test fixture" $
+      parse parseDotGraphs fixtureAcmeFile `shouldSucceedOn` acme
+
+    it "parses package IDs with platforms" $
+      parse parseDotGraphs "" fixturePackageIDWithPlatformContents
+        `shouldParse` fixturePackageIDWithPlatformGraph
+
+    it "renders parsed dependencies" $ do
+      let p =
+            PackageId
+              { groupName = "org.apache.commons"
+              , artifactName = "commons-rng-parent"
+              , artifactType = "pom"
+              , artifactVersion = "1.4-SNAPSHOT"
+              , artifactPlatform = Nothing
+              , buildTag = Nothing
+              }
+      let d =
+            Dependency
+              { dependencyType = MavenType
+              , dependencyName = "org.apache.commons:commons-rng-parent"
+              , dependencyVersion = Just (CEq "1.4-SNAPSHOT")
+              , dependencyLocations = []
+              , dependencyEnvironments = [EnvProduction]
+              , dependencyTags = mempty
+              }
+      toDependency p `shouldBe` d
+
 fixtureSingleFile :: FilePath
 fixtureSingleFile = "test/Maven/testdata/fossa-deptree.dot"
 
@@ -33,6 +74,7 @@ fixtureSingleGraph =
             , artifactName = "commons-rng-simple"
             , artifactType = "jar"
             , artifactVersion = "1.4-SNAPSHOT"
+            , artifactPlatform = Nothing
             , buildTag = Nothing
             }
       , edgeList =
@@ -42,6 +84,7 @@ fixtureSingleGraph =
                 , artifactName = "commons-rng-simple"
                 , artifactType = "jar"
                 , artifactVersion = "1.4-SNAPSHOT"
+                , artifactPlatform = Nothing
                 , buildTag = Nothing
                 }
             , PackageId
@@ -49,6 +92,7 @@ fixtureSingleGraph =
                 , artifactName = "commons-rng-core"
                 , artifactType = "jar"
                 , artifactVersion = "1.4-SNAPSHOT"
+                , artifactPlatform = Nothing
                 , buildTag = Just "compile"
                 }
             )
@@ -58,6 +102,7 @@ fixtureSingleGraph =
                 , artifactName = "commons-rng-simple"
                 , artifactType = "jar"
                 , artifactVersion = "1.4-SNAPSHOT"
+                , artifactPlatform = Nothing
                 , buildTag = Nothing
                 }
             , PackageId
@@ -65,6 +110,7 @@ fixtureSingleGraph =
                 , artifactName = "commons-math3"
                 , artifactType = "jar"
                 , artifactVersion = "3.6.1"
+                , artifactPlatform = Nothing
                 , buildTag = Just "test"
                 }
             )
@@ -74,6 +120,7 @@ fixtureSingleGraph =
                 , artifactName = "commons-rng-simple"
                 , artifactType = "jar"
                 , artifactVersion = "1.4-SNAPSHOT"
+                , artifactPlatform = Nothing
                 , buildTag = Nothing
                 }
             , PackageId
@@ -81,6 +128,7 @@ fixtureSingleGraph =
                 , artifactName = "junit"
                 , artifactType = "jar"
                 , artifactVersion = "4.13.1"
+                , artifactPlatform = Nothing
                 , buildTag = Just "test"
                 }
             )
@@ -90,6 +138,7 @@ fixtureSingleGraph =
                 , artifactName = "commons-rng-core"
                 , artifactType = "jar"
                 , artifactVersion = "1.4-SNAPSHOT"
+                , artifactPlatform = Nothing
                 , buildTag = Just "compile"
                 }
             , PackageId
@@ -97,6 +146,7 @@ fixtureSingleGraph =
                 , artifactName = "commons-rng-client-api"
                 , artifactType = "jar"
                 , artifactVersion = "1.4-SNAPSHOT"
+                , artifactPlatform = Nothing
                 , buildTag = Just "compile"
                 }
             )
@@ -106,6 +156,7 @@ fixtureSingleGraph =
                 , artifactName = "junit"
                 , artifactType = "jar"
                 , artifactVersion = "4.13.1"
+                , artifactPlatform = Nothing
                 , buildTag = Just "test"
                 }
             , PackageId
@@ -113,6 +164,7 @@ fixtureSingleGraph =
                 , artifactName = "hamcrest-core"
                 , artifactType = "jar"
                 , artifactVersion = "1.3"
+                , artifactPlatform = Nothing
                 , buildTag = Just "test"
                 }
             )
@@ -126,101 +178,133 @@ fixtureMultiFile = "test/Maven/testdata/fossa-multideptree.dot"
 fixtureMultiGraph :: [DotGraph]
 fixtureMultiGraph =
   [ DotGraph
-      { rootNode = PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-parent", artifactType = "pom", artifactVersion = "1.4-SNAPSHOT", buildTag = Nothing}
+      { rootNode = PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-parent", artifactType = "pom", artifactVersion = "1.4-SNAPSHOT", artifactPlatform = Nothing, buildTag = Nothing}
       , edgeList =
           [
-            ( PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-parent", artifactType = "pom", artifactVersion = "1.4-SNAPSHOT", buildTag = Nothing}
-            , PackageId{groupName = "junit", artifactName = "junit", artifactType = "jar", artifactVersion = "4.13.1", buildTag = Just "test"}
+            ( PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-parent", artifactType = "pom", artifactVersion = "1.4-SNAPSHOT", artifactPlatform = Nothing, buildTag = Nothing}
+            , PackageId{groupName = "junit", artifactName = "junit", artifactType = "jar", artifactVersion = "4.13.1", artifactPlatform = Nothing, buildTag = Just "test"}
             )
           ,
-            ( PackageId{groupName = "junit", artifactName = "junit", artifactType = "jar", artifactVersion = "4.13.1", buildTag = Just "test"}
-            , PackageId{groupName = "org.hamcrest", artifactName = "hamcrest-core", artifactType = "jar", artifactVersion = "1.3", buildTag = Just "test"}
+            ( PackageId{groupName = "junit", artifactName = "junit", artifactType = "jar", artifactVersion = "4.13.1", artifactPlatform = Nothing, buildTag = Just "test"}
+            , PackageId{groupName = "org.hamcrest", artifactName = "hamcrest-core", artifactType = "jar", artifactVersion = "1.3", artifactPlatform = Nothing, buildTag = Just "test"}
             )
           ]
       }
   , DotGraph
-      { rootNode = PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-client-api", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", buildTag = Nothing}
+      { rootNode = PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-client-api", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", artifactPlatform = Nothing, buildTag = Nothing}
       , edgeList =
-          [ (PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-client-api", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", buildTag = Nothing}, PackageId{groupName = "junit", artifactName = "junit", artifactType = "jar", artifactVersion = "4.13.1", buildTag = Just "test"})
+          [ (PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-client-api", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", artifactPlatform = Nothing, buildTag = Nothing}, PackageId{groupName = "junit", artifactName = "junit", artifactType = "jar", artifactVersion = "4.13.1", artifactPlatform = Nothing, buildTag = Just "test"})
           ,
-            ( PackageId{groupName = "junit", artifactName = "junit", artifactType = "jar", artifactVersion = "4.13.1", buildTag = Just "test"}
-            , PackageId{groupName = "org.hamcrest", artifactName = "hamcrest-core", artifactType = "jar", artifactVersion = "1.3", buildTag = Just "test"}
+            ( PackageId{groupName = "junit", artifactName = "junit", artifactType = "jar", artifactVersion = "4.13.1", artifactPlatform = Nothing, buildTag = Just "test"}
+            , PackageId{groupName = "org.hamcrest", artifactName = "hamcrest-core", artifactType = "jar", artifactVersion = "1.3", artifactPlatform = Nothing, buildTag = Just "test"}
             )
           ]
       }
   , DotGraph
-      { rootNode = PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-core", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", buildTag = Nothing}
+      { rootNode = PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-core", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", artifactPlatform = Nothing, buildTag = Nothing}
       , edgeList =
           [
-            ( PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-core", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", buildTag = Nothing}
-            , PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-client-api", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", buildTag = Just "compile"}
+            ( PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-core", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", artifactPlatform = Nothing, buildTag = Nothing}
+            , PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-client-api", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", artifactPlatform = Nothing, buildTag = Just "compile"}
             )
           ,
-            ( PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-core", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", buildTag = Nothing}
-            , PackageId{groupName = "org.apache.commons", artifactName = "commons-math3", artifactType = "jar", artifactVersion = "3.6.1", buildTag = Just "test"}
+            ( PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-core", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", artifactPlatform = Nothing, buildTag = Nothing}
+            , PackageId{groupName = "org.apache.commons", artifactName = "commons-math3", artifactType = "jar", artifactVersion = "3.6.1", artifactPlatform = Nothing, buildTag = Just "test"}
             )
           ,
-            ( PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-core", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", buildTag = Nothing}
-            , PackageId{groupName = "junit", artifactName = "junit", artifactType = "jar", artifactVersion = "4.13.1", buildTag = Just "test"}
+            ( PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-core", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", artifactPlatform = Nothing, buildTag = Nothing}
+            , PackageId{groupName = "junit", artifactName = "junit", artifactType = "jar", artifactVersion = "4.13.1", artifactPlatform = Nothing, buildTag = Just "test"}
             )
           ,
-            ( PackageId{groupName = "junit", artifactName = "junit", artifactType = "jar", artifactVersion = "4.13.1", buildTag = Just "test"}
-            , PackageId{groupName = "org.hamcrest", artifactName = "hamcrest-core", artifactType = "jar", artifactVersion = "1.3", buildTag = Just "test"}
+            ( PackageId{groupName = "junit", artifactName = "junit", artifactType = "jar", artifactVersion = "4.13.1", artifactPlatform = Nothing, buildTag = Just "test"}
+            , PackageId{groupName = "org.hamcrest", artifactName = "hamcrest-core", artifactType = "jar", artifactVersion = "1.3", artifactPlatform = Nothing, buildTag = Just "test"}
             )
           ]
       }
   , DotGraph
-      { rootNode = PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-simple", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", buildTag = Nothing}
+      { rootNode = PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-simple", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", artifactPlatform = Nothing, buildTag = Nothing}
       , edgeList =
           [
-            ( PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-simple", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", buildTag = Nothing}
-            , PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-core", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", buildTag = Just "compile"}
+            ( PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-simple", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", artifactPlatform = Nothing, buildTag = Nothing}
+            , PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-core", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", artifactPlatform = Nothing, buildTag = Just "compile"}
             )
           ,
-            ( PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-simple", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", buildTag = Nothing}
-            , PackageId{groupName = "org.apache.commons", artifactName = "commons-math3", artifactType = "jar", artifactVersion = "3.6.1", buildTag = Just "test"}
+            ( PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-simple", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", artifactPlatform = Nothing, buildTag = Nothing}
+            , PackageId{groupName = "org.apache.commons", artifactName = "commons-math3", artifactType = "jar", artifactVersion = "3.6.1", artifactPlatform = Nothing, buildTag = Just "test"}
             )
           ,
-            ( PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-simple", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", buildTag = Nothing}
-            , PackageId{groupName = "junit", artifactName = "junit", artifactType = "jar", artifactVersion = "4.13.1", buildTag = Just "test"}
+            ( PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-simple", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", artifactPlatform = Nothing, buildTag = Nothing}
+            , PackageId{groupName = "junit", artifactName = "junit", artifactType = "jar", artifactVersion = "4.13.1", artifactPlatform = Nothing, buildTag = Just "test"}
             )
           ,
-            ( PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-core", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", buildTag = Just "compile"}
-            , PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-client-api", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", buildTag = Just "compile"}
+            ( PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-core", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", artifactPlatform = Nothing, buildTag = Just "compile"}
+            , PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-client-api", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", artifactPlatform = Nothing, buildTag = Just "compile"}
             )
           ,
-            ( PackageId{groupName = "junit", artifactName = "junit", artifactType = "jar", artifactVersion = "4.13.1", buildTag = Just "test"}
-            , PackageId{groupName = "org.hamcrest", artifactName = "hamcrest-core", artifactType = "jar", artifactVersion = "1.3", buildTag = Just "test"}
+            ( PackageId{groupName = "junit", artifactName = "junit", artifactType = "jar", artifactVersion = "4.13.1", artifactPlatform = Nothing, buildTag = Just "test"}
+            , PackageId{groupName = "org.hamcrest", artifactName = "hamcrest-core", artifactType = "jar", artifactVersion = "1.3", artifactPlatform = Nothing, buildTag = Just "test"}
             )
           ]
       }
   , DotGraph
-      { rootNode = PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-sampling", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", buildTag = Nothing}
+      { rootNode = PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-sampling", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", artifactPlatform = Nothing, buildTag = Nothing}
       , edgeList =
           [
-            ( PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-sampling", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", buildTag = Nothing}
-            , PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-client-api", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", buildTag = Just "compile"}
+            ( PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-sampling", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", artifactPlatform = Nothing, buildTag = Nothing}
+            , PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-client-api", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", artifactPlatform = Nothing, buildTag = Just "compile"}
             )
           ,
-            ( PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-sampling", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", buildTag = Nothing}
-            , PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-simple", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", buildTag = Just "test"}
+            ( PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-sampling", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", artifactPlatform = Nothing, buildTag = Nothing}
+            , PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-simple", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", artifactPlatform = Nothing, buildTag = Just "test"}
             )
           ,
-            ( PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-sampling", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", buildTag = Nothing}
-            , PackageId{groupName = "org.apache.commons", artifactName = "commons-math3", artifactType = "jar", artifactVersion = "3.6.1", buildTag = Just "test"}
+            ( PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-sampling", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", artifactPlatform = Nothing, buildTag = Nothing}
+            , PackageId{groupName = "org.apache.commons", artifactName = "commons-math3", artifactType = "jar", artifactVersion = "3.6.1", artifactPlatform = Nothing, buildTag = Just "test"}
             )
           ,
-            ( PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-sampling", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", buildTag = Nothing}
-            , PackageId{groupName = "junit", artifactName = "junit", artifactType = "jar", artifactVersion = "4.13.1", buildTag = Just "test"}
+            ( PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-sampling", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", artifactPlatform = Nothing, buildTag = Nothing}
+            , PackageId{groupName = "junit", artifactName = "junit", artifactType = "jar", artifactVersion = "4.13.1", artifactPlatform = Nothing, buildTag = Just "test"}
             )
           ,
-            ( PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-simple", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", buildTag = Just "test"}
-            , PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-core", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", buildTag = Just "test"}
+            ( PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-simple", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", artifactPlatform = Nothing, buildTag = Just "test"}
+            , PackageId{groupName = "org.apache.commons", artifactName = "commons-rng-core", artifactType = "jar", artifactVersion = "1.4-SNAPSHOT", artifactPlatform = Nothing, buildTag = Just "test"}
             )
           ,
-            ( PackageId{groupName = "junit", artifactName = "junit", artifactType = "jar", artifactVersion = "4.13.1", buildTag = Just "test"}
-            , PackageId{groupName = "org.hamcrest", artifactName = "hamcrest-core", artifactType = "jar", artifactVersion = "1.3", buildTag = Just "test"}
+            ( PackageId{groupName = "junit", artifactName = "junit", artifactType = "jar", artifactVersion = "4.13.1", artifactPlatform = Nothing, buildTag = Just "test"}
+            , PackageId{groupName = "org.hamcrest", artifactName = "hamcrest-core", artifactType = "jar", artifactVersion = "1.3", artifactPlatform = Nothing, buildTag = Just "test"}
             )
           ]
       }
   ]
+
+fixtureAcmeFile :: FilePath
+fixtureAcmeFile = "test/Maven/testdata/acme-deptree.dot"
+
+fixturePackageIDWithPlatformContents :: Text
+fixturePackageIDWithPlatformContents =
+  "digraph \"com.fossa:fixture:pom:1.2.3\" {\
+  \  \"com.fossa:fixture:pom:1.2.3\" -> \"com.fossa:platform-specific-dep:jar:linux-x86_64:4.5.6:compile\" ;\
+  \}"
+
+fixturePackageIDWithPlatformGraph :: [DotGraph]
+fixturePackageIDWithPlatformGraph =
+  [DotGraph{rootNode = root, edgeList = [(root, platformID)]}]
+  where
+    root =
+      PackageId
+        { groupName = "com.fossa"
+        , artifactName = "fixture"
+        , artifactType = "pom"
+        , artifactVersion = "1.2.3"
+        , artifactPlatform = Nothing
+        , buildTag = Nothing
+        }
+    platformID =
+      PackageId
+        { groupName = "com.fossa"
+        , artifactName = "platform-specific-dep"
+        , artifactType = "jar"
+        , artifactVersion = "4.5.6"
+        , artifactPlatform = Just "linux-x86_64"
+        , buildTag = Just "compile"
+        }